Added unit coordinate searching

+++ Scadue.Data/Models/UnitCoordinatesEntity.cs
@@ -2,8 +2,8 @@
 {
     public class UnitCoordinatesEntity : Entity
     {
-        public float Lat { get; set; }
-        public float Lon { get; set; }
+        public double Lat { get; set; }
+        public double Lon { get; set; }
         public int Number { get; set; }
         public int UnitId { get; set; }
         public AdministrativeUnitEntity AdministrativeUnit { get; set; }
